progress: move into the final place in images and fix imports
This commit finishes moving the bootc-image-builder "progress"
package into "images". This allows code-sharing between bib
and ibcli.

Note that further work will happen to unify `osbuild.RunOsbuild`
and `progress.RunOSBuild()`.
